requirements.txt
appnope==0.1.2
backcall==0.2.0
bleach==4.1.0
certifi==2021.5.30
charset-normalizer==2.0.6
decorator==5.1.0
idna==3.2
ipdb==0.13.9
ipython==7.28.0
jedi==0.18.0
Jinja2==3.0.1
Markdown==3.3.4
markdown-checklist==0.4.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mdx-linkify==2.1
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
PyYAML==5.4.1
raylib-py==0.1.1
requests==2.26.0
six==1.16.0
toml==0.10.2
traitlets==5.1.0
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
